Fix sast.java.directory_traversal in src/main/java/org/owasp/webgoat/lessons/challenges/challenge7/MD5.java

+++ src/main/java/org/owasp/webgoat/webwolf/FileServer.java
@@ -12,6 +12,10 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.nio.file.attribute.FileTime;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
@@ -72,7 +76,11 @@ public class FileServer {
     // DO NOT use multipartFile.transferTo(), see
     // https://stackoverflow.com/questions/60336929/java-nio-file-nosuchfileexception-when-file-transferto-is-called
     try (InputStream is = multipartFile.getInputStream()) {
-      var destinationFile = destinationDir.toPath().resolve(multipartFile.getOriginalFilename());
+      Path basePath = Paths.get(fileLocation, username).toAbsolutePath().normalize();
+      Path destinationFile = basePath.resolve(multipartFile.getOriginalFilename()).normalize();
+      if (!destinationFile.startsWith(basePath)) {
+        throw new SecurityException("Invalid file path");
+      }
       Files.deleteIfExists(destinationFile);
       Files.copy(is, destinationFile);
     }
@@ -87,11 +95,16 @@ public class FileServer {
   public ModelAndView getFiles(
       HttpServletRequest request, Authentication authentication, TimeZone timezone) {
     String username = (null != authentication) ? authentication.getName() : "anonymous";
-    File destinationDir = new File(fileLocation, username);
+    File destinationDir = new File(fileLocation.getPath(), username);
 
     ModelAndView modelAndView = new ModelAndView();
     modelAndView.setViewName("files");
-    File changeIndicatorFile = new File(destinationDir, username + "_changed");
+    Path basePath = Paths.get(fileLocation).toAbsolutePath().normalize();
+    Path changeIndicatorPath = Paths.get(destinationDir, username + "_changed").toAbsolutePath().normalize();
+    if (!changeIndicatorPath.startsWith(basePath)) {
+      throw new SecurityException("Access denied to change indicator file");
+    }
+    File changeIndicatorFile = changeIndicatorPath.toFile();
     if (changeIndicatorFile.exists()) {
       modelAndView.addObject("uploadSuccess", request.getParameter("uploadSuccess"));
     }
